% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{asModule}
\alias{asModule}
\title{Wrap a component as a Shiny module}
\usage{
asModule(component)
}
\arguments{
\item{component}{A component list returned by \code{\link[=component]{component()}}.}
}
\value{
A list with \code{ui} and \code{server} functions for \code{\link[shiny:moduleServer]{shiny::moduleServer()}}.
}
\description{
Wrap a component as a Shiny module
}
