% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\name{effect}
\alias{effect}
\title{Declarative effect specification for \code{\link[=component]{component()}}}
\usage{
effect(deps = NULL, fn)
}
\arguments{
\item{deps}{Character vector of state field names to watch, or \code{NULL} to run
once on mount.}

\item{fn}{Function invoked when dependencies change. Receives the state
accessor.}
}
\description{
Declarative effect specification for \code{\link[=component]{component()}}
}
