% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Store.R
\name{getStore}
\alias{getStore}
\title{Get a store in the given Shiny session.}
\usage{
getStore(name, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{name}{The name of the store.}

\item{session}{The Shiny session to retrieve the store from.}
}
\value{
A \code{\link{Store}} object.
}
\description{
Gets an existing store for the given session.
}
\examples{
states <- list(data = "mtcars")
events <- list("data_changed")
session <- shiny::MockShinySession$new()
new_store(name = "app", states = states, events = events, session = session)
# in a module server:
store <- getStore("app", session)
}
