% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{mount}
\alias{mount}
\title{Mount a component in a Shiny UI}
\usage{
mount(component, ...)
}
\arguments{
\item{component}{A component list returned by \code{\link[=component]{component()}}.}

\item{...}{Passed to the component UI function.}
}
\description{
Mount a component in a Shiny UI
}
