% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Store.R
\name{new_store}
\alias{new_store}
\title{Create a new store for the given Shiny session.}
\usage{
new_store(
  name,
  states,
  events = NULL,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{name}{The name of the store.}

\item{states}{A named list of states and their values.}

\item{events}{An optional list of event names to manage with the store.}

\item{session}{The Shiny session to save the store in.}
}
\value{
A \code{\link{Store}} object.
}
\description{
Creates a \code{\link{Store}} object for the given session. Store names
must be unique within a session. The store is created in the \code{userData}
environment of the given session.
}
\examples{
states <- list(data = "mtcars")
events <- list("data_changed")
session <- shiny::MockShinySession$new()
store <- new_store(name = "app", states = states, events = events, session = session)
store$getState("data")

}
