% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{saveInterfaceLocal}
\alias{saveInterfaceLocal}
\alias{loadInterfaceLocal}
\title{Set saving and loading interfaces for shiny bookmarking local storage}
\usage{
saveInterfaceLocal(id, callback)

loadInterfaceLocal(id, callback)
}
\arguments{
\item{id}{character string for session ID.}

\item{callback}{function to call with the path to save/load the bookmark state.}
}
\description{
The \code{saveInterfaceLocal} and \code{loadInterfaceLocal} functions provide
implementations for saving and loading Shiny bookmark state to a
local directory. These functions can be set as the saving and
loading interfaces for Shiny bookmarking using
\code{shiny::shinyOptions()}.. While these callback functions are
set by default when initializing a new instance of \code{StorageClass},
certain Shiny application structures such as applications created
with the \code{golem} R package require these callbacks to be defined
as part of the \code{onStart} argument in \code{shiny::runApp()}.
}
