% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{serve}
\alias{serve}
\title{Run server logic for a mounted component}
\usage{
serve(component, input, output, session)
}
\arguments{
\item{component}{A component list returned by \code{\link[=component]{component()}}.}

\item{input, output, session}{Shiny server function arguments.}
}
\description{
Run server logic for a mounted component
}
