% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state-accessor.R
\name{shinystate_state}
\alias{shinystate_state}
\title{Component state accessor}
\format{
An environment with S3 access to state fields.
}
\description{
Object returned inside a component \code{render()} function for reading and
updating state. Use \verb{$field} to read and \verb{$set()} to write.
}
