% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Store.R
\name{Store}
\alias{Store}
\title{R6 Class Implementing a Store}
\description{
A \code{Store} is an R6 object for storing and managing the state of
a Shiny application. This includes event handling.
}
\examples{
states <- list(data = "mtcars")
events <- list("data_changed")
store <- Store$new(name = "app", states = states, events = events)
store$getState("data")
store$setState("data", "iris")
store$getState("data")

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Store-new}{\code{Store$new()}}
\item \href{#method-Store-getState}{\code{Store$getState()}}
\item \href{#method-Store-setState}{\code{Store$setState()}}
\item \href{#method-Store-trigger}{\code{Store$trigger()}}
\item \href{#method-Store-on}{\code{Store$on()}}
\item \href{#method-Store-clone}{\code{Store$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Store-new"></a>}}
\if{latex}{\out{\hypertarget{method-Store-new}{}}}
\subsection{Method \code{new()}}{
Initialize a \code{Store} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Store$new(name, states, events = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the store.}

\item{\code{states}}{A named list of states and their values.}

\item{\code{events}}{A list of event names to manage with the store.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Store-getState"></a>}}
\if{latex}{\out{\hypertarget{method-Store-getState}{}}}
\subsection{Method \code{getState()}}{
Get the value of a state.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Store$getState(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the state.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The value of the state.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Store-setState"></a>}}
\if{latex}{\out{\hypertarget{method-Store-setState}{}}}
\subsection{Method \code{setState()}}{
Set the value of a state.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Store$setState(name, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the state.}

\item{\code{value}}{The value of the state.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Self invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Store-trigger"></a>}}
\if{latex}{\out{\hypertarget{method-Store-trigger}{}}}
\subsection{Method \code{trigger()}}{
Trigger an event.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Store$trigger(event)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{event}}{The name of the event.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Store-on"></a>}}
\if{latex}{\out{\hypertarget{method-Store-on}{}}}
\subsection{Method \code{on()}}{
Respond to an event.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Store$on(event, expression)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{event}}{The name of the event.}

\item{\code{expression}}{The expression to run when the event occurs.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Store-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Store-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Store$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
