% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{Component}
\alias{Component}
\title{Create components}
\usage{
Component(..., template, export)
}
\arguments{
\item{...}{Values to load into the component environment or functions.
Functions arguments may default to reactives in which they too are
reactive.}

\item{template}{A tag element.}

\item{export}{A call to \code{exports()}, currently unused.}
}
\description{
A component connects a front-end \code{template} to exports (WIP).
}
\examples{
library(shiny)

Component(
  template = actionButton(
    inputId = "clicks",
    label = "A button"
  ),
  function(clicks = input$clicks) {
    print("Button clicked")
  }
)

}
