% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinystate-package.R
\docType{package}
\name{shinystate-package}
\alias{shinystate}
\alias{shinystate-package}
\title{shinystate: Shiny, Stateful, Modular}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
