% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storageclass.R
\name{StorageClass}
\alias{StorageClass}
\title{StorageClass R6 class}
\description{
This class provides a set of methods to create a Shiny bookmarkable state
storage location using R6.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StorageClass-new}{\code{StorageClass$new()}}
\item \href{#method-StorageClass-restore}{\code{StorageClass$restore()}}
\item \href{#method-StorageClass-snapshot}{\code{StorageClass$snapshot()}}
\item \href{#method-StorageClass-register_metadata}{\code{StorageClass$register_metadata()}}
\item \href{#method-StorageClass-clone}{\code{StorageClass$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageClass-new"></a>}}
\if{latex}{\out{\hypertarget{method-StorageClass-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageClass$new(local_storage_dir = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{local_storage_dir}}{file path to use for storing bookmarkable state
files. If not specified, a temporary directory on the host system
will be used.}

\item{\code{bmi_storage}}{TODO may move to private}

\item{\code{board_sessions}}{TODO may move to private}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Create a new 'StorageClass' object.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageClass-restore"></a>}}
\if{latex}{\out{\hypertarget{method-StorageClass-restore}{}}}
\subsection{Method \code{restore()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageClass$restore(url, session = shiny::getDefaultReactiveDomain())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{character with the unique URL assigned to the bookmarkable
state session.}

\item{\code{session}}{The Shiny session to associate with the restore operation}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Restore a previous bookmarkable state session
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageClass-snapshot"></a>}}
\if{latex}{\out{\hypertarget{method-StorageClass-snapshot}{}}}
\subsection{Method \code{snapshot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageClass$snapshot(
  session_metadata = NULL,
  session = shiny::getDefaultReactiveDomain()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{session_metadata}}{Optional named list of additional variables to
include with the default bookmarkable state attributes when creating
the snapshot. Each element of the list must be a single-length item}

\item{\code{session}}{The Shiny session to associate with the snapshot operation}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Create a snapshot of bookmarkable state
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageClass-register_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-StorageClass-register_metadata}{}}}
\subsection{Method \code{register_metadata()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageClass$register_metadata()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Register bookmarkable state storage data collection
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageClass-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StorageClass-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
